requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.1            # via django
attrs==20.3.0             # via pytest
certifi==2020.12.5       # via requests
chardet==4.0.0            # via requests
django==3.1.5             # via -r requirements.in, djangorestframework
django-filter==2.4.0
djangorestframework==3.12.2  # via -r requirements.in
idna==3.1                 # via requests
importlib-metadata==3.4.0  # via pluggy, pytest
more-itertools==8.6.0     # via pytest
packaging==20.8           # via pytest
pluggy==0.13.1            # via pytest
py==1.10.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==4.1.0      # via -r requirements.in
pytest==6.2.1             # via pytest-django
pytz==2020.5              # via django
requests==2.25.1          # via -r requirements.in
six==1.15.0               # via packaging
sqlparse==0.4.1           # via django
urllib3==1.26.2           # via requests
wcwidth==0.2.5            # via pytest
zipp==3.4.0               # via importlib-metadata
django-cors-headers==3.6.0
